Add files via upload
Project Methodology: 
-Data Cleaning & Pre-processing 
-Feature Engineering 
-Splitting data into train/test/validation sets -Model training and testing 
-Model evaluation

Model Evaluation: 
The final modelchosen is ‘Neural Network’model with dropout 
layers. It works efficiently for the task of classification upon our 
training set after being trained and validated. Also, it provides us 
with an accuracy of almost 70% upon using the test set, which 
could not be achieved while testing various decision tree or 
logistic regression models with different parameters. Dropout is a 
regularization technique that can be used to reduce overfitting in 
neural networks. The idea is to randomly drop out (i.e., set to zero) 
some fraction of the neurons in each hidden layer during training. 
This helps to prevent the network from relying too much on any 
one neuron or set of neurons and encourages the network to learn more robust and generalizable features.
confusion Matrics Generated to indicate that our Model predictions for a total of 20 samples in the test set. Out of these, 5 were 
true negatives (TN), 4 were false positives (FP), 2 were false negatives 
(FN), and 9 were true positives (TP). This depicts a 70% accuracy of our model which means that if this model was used for store performance 
classification, 70% of the stores would be correctly predicted.